Restructure packaging for PayByBank
COAND-497
