+++ paybybank/src/main/java/com/adyen/checkout/paybybank/internal/ui/PayByBankDelegate.kt
@@ -6,7 +6,7 @@
  * Created by ozgur on 28/9/2022.
  */
 
-package com.adyen.checkout.paybybank
+package com.adyen.checkout.paybybank.internal.ui
 
 import com.adyen.checkout.components.PaymentComponentState
 import com.adyen.checkout.components.base.PaymentComponentDelegate
@@ -14,9 +14,11 @@ import com.adyen.checkout.components.model.payments.request.PayByBankPaymentMeth
 import com.adyen.checkout.components.ui.UIStateDelegate
 import com.adyen.checkout.components.ui.ViewProvidingDelegate
 import com.adyen.checkout.issuerlist.internal.ui.model.IssuerModel
+import com.adyen.checkout.paybybank.internal.ui.model.PayByBankInputData
+import com.adyen.checkout.paybybank.internal.ui.model.PayByBankOutputData
 import kotlinx.coroutines.flow.Flow
 
-interface PayByBankDelegate :
+internal interface PayByBankDelegate :
     PaymentComponentDelegate<PaymentComponentState<PayByBankPaymentMethod>>,
     ViewProvidingDelegate,
     UIStateDelegate {
