+++ paybybank/src/main/java/com/adyen/checkout/paybybank/internal/ui/model/PayByBankInputData.kt
@@ -6,12 +6,12 @@
  * Created by ozgur on 27/9/2022.
  */
 
-package com.adyen.checkout.paybybank
+package com.adyen.checkout.paybybank.internal.ui.model
 
 import com.adyen.checkout.components.base.InputData
 import com.adyen.checkout.issuerlist.internal.ui.model.IssuerModel
 
-class PayByBankInputData(
+internal class PayByBankInputData(
     var query: String? = null,
     var selectedIssuer: IssuerModel? = null
 ) : InputData
