+++ paybybank/src/main/java/com/adyen/checkout/paybybank/internal/ui/model/PayByBankOutputData.kt
@@ -6,12 +6,12 @@
  * Created by ozgur on 27/9/2022.
  */
 
-package com.adyen.checkout.paybybank
+package com.adyen.checkout.paybybank.internal.ui.model
 
 import com.adyen.checkout.components.base.OutputData
 import com.adyen.checkout.issuerlist.internal.ui.model.IssuerModel
 
-data class PayByBankOutputData(
+internal data class PayByBankOutputData(
     val selectedIssuer: IssuerModel?,
     val issuers: List<IssuerModel>,
 ) : OutputData {
