+++ paybybank/src/test/java/com/adyen/checkout/paybybank/PayByBankComponentTest.kt
@@ -19,6 +19,8 @@ import com.adyen.checkout.components.base.ComponentEventHandler
 import com.adyen.checkout.components.model.payments.request.PayByBankPaymentMethod
 import com.adyen.checkout.components.test.TestComponentViewType
 import com.adyen.checkout.core.log.Logger
+import com.adyen.checkout.paybybank.internal.ui.PayByBankComponentViewType
+import com.adyen.checkout.paybybank.internal.ui.PayByBankDelegate
 import com.adyen.checkout.test.TestDispatcherExtension
 import com.adyen.checkout.test.extensions.invokeOnCleared
 import kotlinx.coroutines.ExperimentalCoroutinesApi
